test: split activeSessionSchema tests into a separate file
A static import of the router module in the listInstances test evaluated
config.server.ts (freezing SESSION_INGEST_WORKER_URL at the empty .env.test
value) before beforeAll could set the env, breaking the dynamic-import dance.
Move the schema-validation cases to active-sessions-router.schema.test.ts,
which Jest sandboxes separately.
