+++ apps/web/src/routers/active-sessions-router.test.ts
@@ -1,89 +1,17 @@
 import { describe, expect, it, jest, beforeAll, afterEach } from '@jest/globals';
 import { TRPCError } from '@trpc/server';
-import { activeSessionSchema } from './active-sessions-router';
 import { insertTestUser } from '@/tests/helpers/user.helper';
 import { db } from '@/lib/drizzle';
 import { organizations, organization_memberships } from '@kilocode/db/schema';
 import type { User, Organization } from '@kilocode/db/schema';
 import type { createCallerForUser as CreateCallerForUser } from '@/routers/test-utils';
 
-describe('activeSessionSchema capabilities', () => {
-  it('accepts a session row with capabilities.attachments: true', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-      capabilities: { attachments: true },
-    };
-    expect(activeSessionSchema.safeParse(row).success).toBe(true);
-  });
-
-  it('accepts a session row with capabilities.attachments: false', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-      capabilities: { attachments: false },
-    };
-    expect(activeSessionSchema.safeParse(row).success).toBe(true);
-  });
-
-  it('accepts a session row with an empty capabilities object', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-      capabilities: {},
-    };
-    expect(activeSessionSchema.safeParse(row).success).toBe(true);
-  });
-
-  it('accepts a session row with an absent capabilities field (legacy CLI)', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-    };
-    const result = activeSessionSchema.safeParse(row);
-    expect(result.success).toBe(true);
-    if (result.success) {
-      expect(result.data.capabilities).toBeUndefined();
-    }
-  });
-
-  it('rejects a non-boolean capabilities.attachments value', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-      capabilities: { attachments: 'yes' },
-    };
-    expect(activeSessionSchema.safeParse(row).success).toBe(false);
-  });
-
-  it('rejects unknown capability keys (strict object)', () => {
-    const row = {
-      id: 's1',
-      status: 'busy',
-      title: 'Fix bug',
-      connectionId: 'cli-1',
-      capabilities: { terminal: true },
-    };
-    // The default zod behavior strips unknown keys rather than rejecting —
-    // assert that the unknown key is dropped, not preserved, so consumers
-    // never see a flag the cloud did not advertise.
-    const result = activeSessionSchema.safeParse(row);
-    expect(result.success).toBe(true);
-    if (result.success) {
-      expect(result.data.capabilities).toEqual({});
-    }
-  });
-});
+// NOTE: `activeSessionSchema` validation is covered in the sibling
+// `active-sessions-router.schema.test.ts`. It is deliberately NOT imported
+// here: a static import of the router module evaluates `config.server.ts`
+// (and freezes `SESSION_INGEST_WORKER_URL` at the empty `.env.test` value)
+// before the `beforeAll` below can set the env — which is exactly what the
+// dynamic-import dance is designed to avoid.
 
 // `.env.test` sets SESSION_INGEST_WORKER_URL to '' (shared fixture used by
 // other test files too — do not change it here). `createCallerForUser`'s
